+++ src/pages/demo/partials/personProfiles.tsx
@@ -4,6 +4,7 @@ import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
 import Container from '@/components/container'
 
 const PersonProfilePage = ({ nodes }) => {
+  console.log(nodes)
   if (!nodes) return
   return (
     <>
